# doc/api.yaml
openapi: 3.0.0
info:
  description: |-
    This concerns the "Web and Software Architecture" course to develop from scratch a web application that allows users to share photos of special moments.
    It allows users to like, comment on posts, ban other users, and so forth.

    Note: "Banning" and "setting a profile to private" are different. Setting a profile to public (private = false), means that any user except those whom he bans can see the user's profile, comment, like, etc without following the user. This is a feature much like "explore" in many social media. If a user wants only his followers to see his feeds except those banned, he should set it to private. Otherwise, even those who did not follow him and were not banned by him can see his feeds.
  title: Photo Sharing App
  version: 1.0.0
  contact:
    name: Yusupha Juwara
    url: https://github.com/YusuphaJuwara
    email: yusuphajuwara@gmail.com
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  parameters:
...
    userID:  #Arbitrary name for the definition that will be used to refer to it. Not necessarily the same as the parameter name.
      in: path
      name: user-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: The user ID
    postID:
      in: path
      name: post-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: The post ID
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: uuid    # optional, arbitrary value for documentation purposes
  responses:
    NoContent204:
      description: Successful.
    BadRequest400:
      description: |-
        The request was not compliant with the documentation (eg. missing fields, etc).
        Further info in server logs.
    UnAuth401:
      description: |-
        Access token is missing or invalid.
        The request was not authorized. The user must be authorized first to perform this request.
    Forbidden403:
      description: |-
        Trying to access a resource that the user should not. This is forbidden action.
        E.g., 
        a banned user trying to get an info about the user who banned him, 
        or trying to delete a comment when the user is not the one who placed the comment or on whose post the comment was placed, etc.
        Further info in server logs.
    NotFound404:
      description: Not found.
    InternalServerError500:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  schemas:
    ID:
      description: The unique identifier. It is 36 characters containing 4 "-". 
      type: string
      format: uuid

      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 123e4567-e89b-12d3-a456-426655440000
...
    date-time:
      description: The date and time the post/comment was posted.
      type: string
      format: date-time
      example: '2017-07-21T17:32:28Z'

      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z$'
      minLength: 20   
      maxLength: 20
...
    message:
      description: The message as a comment or caption on a post.
      type: string
      pattern: '(?=.*?[a-zA-Z])[^]'  # Any char including new lines of at least one alphabet.
      minLength: 2
      maxLength: 1000
      example: "Wow, that's a beautifull family photo!!!"
    hashtag:
      description: |-
        Hashtag contains at least one alphabet and zero or more other alphanumeric characters in [a-zA-Z0-9].
      type: string
      pattern: '^(?=.*?[a-zA-Z])[a-zA-Z0-9]{1,20}$'
      minLength: 1
      maxLength: 20
      example: "covid19"
    hashtags:
      description: Every post also includes the hashtags
      type: array
      minItems: 0
      maxItems: 100
      items:  
        $ref: '#/components/schemas/hashtag'
    count:
      description: The total count of the object such as total number of likes, total number of comments, etc.
      type: integer
      minimum: 0
      example: 1000
...
    photo:
      description: The photo in a PNG format. The length is in Bytes.
      type: string
      format: binary

      minLength: 1
      maxLength: 1000000
...
    post: 
      type: object
      description: details related to post.
      properties:
        post-id:  
          $ref: '#/components/schemas/ID'
        photo-id:   
          $ref: '#/components/schemas/ID'
        user-id:  
          $ref: '#/components/schemas/ID'
        date-time:  
          $ref: '#/components/schemas/date-time'
        caption:  
          $ref: '#/components/schemas/message'
        like-count:  
          $ref: '#/components/schemas/count'
        comment-count:   
          $ref: '#/components/schemas/count'
        hashtags: 
            $ref: '#/components/schemas/hashtags'

# 2) Apply security globally to all operations except log-in
security:
  - bearerToken: []
...
paths:
...
  /posts/{post-id}:
    parameters:
      - $ref: "#/components/parameters/postID"
    get:
      tags: [Post]
      #I prefer calling it getPost because I upload a photo with hashtags, captions, etc instead of just a photo. 
      #So what the prof refers to as photo, is my post (photo with hashtags, captions, etc). 
      #But following the project's specs, I will use the operationId keys specified in the project's specification.
      operationId: getPhoto
      summary: Get the post, given the post ID.
      description: |-
        Given the post ID, this sends the post with all its metadata attached. All the metatdata such as the message, the likes, comments, an ID to retrieve the photo, and the date and time posted, and the hashtags.
      responses:
        '200':
          description: The requested post
          content:
            application/json:
              schema:   
                $ref: '#/components/schemas/post'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /posts/:
    get:
      tags: [Post]
      operationId: getPhotos
      summary: Get all posts of all users who did not set their profiles to private and did not ban the user.
      description: |-
        This returns a list of posts of users who did not set their profiles to private and did not ban the user. 
        This is similar to \"explore\" in some Apps, where everyone can see posts that are public (or were not banned by the users who posted them). Then may want to follow users they did not follow already, like the posts, comment on them, etc. 
        Note that this is different from the /users/{user-id}/posts/stream/ defined below 
      responses:
        '200':
          description: Returns an array of posts. The fields are defined in the post section
          content:
            application/json:
              schema: 
                description: An array of posts
                type: array
                minItems: 0
                maxItems: 1000
                items:  
                  $ref: '#/components/schemas/post'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /users/{user-id}/posts/stream/:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [Post]
      operationId: getMyStream
      summary: Get the list of posts posted by the given user's followings.
      description: |-
        Given a user ID, returns an array of posts posted by the given user's followings. This does not include what the user himself posts (read the requirement of the project to understand). What a post returns, is described in the post section.
      responses:
        '200':
          description: An array of Requested posts of the user followings.
          content:
            application/json:
              schema: 
                description: An array of Posts of the user's followings.
                type: array
                minItems: 0
                maxItems:  1000
                items:  
                  $ref: '#/components/schemas/post'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /users/{user-id}/posts/:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [Post]
      #I prefer calling it getUserPosts because I get a photo with hashtags, captions, date and time posted, etc instead of just a photo. 
      #So what the prof refers to as photo, is my post (photo with hashtags, captions, etc). 
      #But following the project's specs, I will use the operationId keys specified in the project's specification.
      operationId: getUserPhotos
      summary: Get the list of posts posted by the given user.
      description: |-
        Given a user ID, returns an array of posts posted by the given user himself. What a post returns, is described in the post section.
      responses:
        '200':
          description: Requested posts of the user.
          content:
            application/json:
              schema: 
                description: An array of posts of the given user.
                type: array
                minItems: 0
                maxItems: 1000
                items:  
                  $ref: '#/components/schemas/post'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    post:
      tags: [Post]
      #I prefer calling it uploadPost because I upload a photo with hashtags, captions, etc instead of just a photo. 
      #So what the prof refers to as photo, is my post (photo with hashtags, captions, etc). 
      #But following the project's specs, I will use the operationId keys specified in the project's specification.
      operationId: uploadPhoto
      summary: Create a new Post by the given user and returns the newly created post ID.
      description: |-
        Create a new \"post\" with a photo that must be present and optionally a caption and an array of hashtags.
        Caption and hashtags can even be added later on using another endpoint.
        The server will create a new unique ID, the client can find it in the response.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: 
              description: |-
                An object containing 2 other objects (photo and caption) and an array of hashtags.
              type: object
              properties:
                photo:  
                  $ref: '#/components/schemas/photo'
                caption:  
                  $ref: '#/components/schemas/message'
                hashtags:  
                    $ref: '#/components/schemas/hashtags'
              required:
                - photo
      responses:
        "201":
          description: Post created successfully
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/ID' 
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /users/{user-id}/posts/{post-id}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/postID"
    put:
      tags: [Post]
      operationId: modifyCaption
      summary: Replaces the caption associated with the post with the new caption.
      description: |-
        Given a new caption, it replaces the old one with the new one.
        Note that this method replaces only the caption and not the photo. 
        The photo cannot be replaced unless the post is deleted and a new one with the intended photo is posted again.
      requestBody:
        required: true
        content:
          application/json:
            schema:   
              $ref: '#/components/schemas/message'
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500" 
    delete:
      tags: [Post]
      operationId: deletePhoto  #I prefer deletePost
      summary: Deletes the post including the photo, caption, likes and comments, date and time posted, etc.
      description: |-
        Deletes the post with the given post ID together with the photo, caption, likes and comments, etc.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
...